Switched lures and rods to php
I set all links to the lures and rods pages to the php version instead of the html. Also had to make slight changes to the php files so they would link to the html files. Make sure any changes made to the rods and lures pages are done to the PHP and NOT the HTML.